requirements.txt
Flask>=1.0.2
Keras>=2.2.4
coverage>=4.5.2
gensim>=3.6.0
matplotlib>=3.0.2
nltk>=3.4
pyenchant>=2.0.0
pytest>=4.0.0
scikit-learn>=0.20.0
seaborn>=0.9.0
spacy>=2.0.16
tensorflow-gpu>=1.12.0